+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsinterventionsservice/service/notifications/NotifyService.kt
@@ -234,48 +234,3 @@ class NotifyAppointmentService(
     }
   }
 }
-
-@Service
-class NotifyReferralService(
-  @Value("\${notify.templates.referral-sent}") private val referralSentTemplateID: String,
-  @Value("\${notify.templates.referral-assigned}") private val referralAssignedTemplateID: String,
-  @Value("\${interventions-ui.baseurl}") private val interventionsUIBaseURL: String,
-  @Value("\${interventions-ui.locations.service-provider.referral-details}") private val spReferralDetailsLocation: String,
-  private val emailSender: EmailSender,
-  private val hmppsAuthService: HMPPSAuthService,
-) : ApplicationListener<ReferralEvent>, NotifyService {
-
-  @AsyncEventExceptionHandling
-  override fun onApplicationEvent(event: ReferralEvent) {
-    when (event.type) {
-      ReferralEventType.SENT -> {
-        val location = generateResourceUrl(interventionsUIBaseURL, spReferralDetailsLocation, event.referral.id)
-        val intervention = event.referral.intervention
-        val serviceProvider = intervention.dynamicFrameworkContract.primeProvider
-        emailSender.sendEmail(
-          referralSentTemplateID,
-          intervention.incomingReferralDistributionEmail,
-          mapOf(
-            "organisationName" to serviceProvider.name,
-            "referenceNumber" to event.referral.referenceNumber!!,
-            "referralUrl" to location.toString(),
-          )
-        )
-      }
-
-      ReferralEventType.ASSIGNED -> {
-        val userDetails = hmppsAuthService.getUserDetail(event.referral.currentAssignee!!)
-        val location = generateResourceUrl(interventionsUIBaseURL, spReferralDetailsLocation, event.referral.id)
-        emailSender.sendEmail(
-          referralAssignedTemplateID,
-          userDetails.email,
-          mapOf(
-            "spFirstName" to userDetails.firstName,
-            "referenceNumber" to event.referral.referenceNumber!!,
-            "referralUrl" to location.toString(),
-          )
-        )
-      }
-    }
-  }
-}
