+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsinterventionsservice/service/notifications/ReferralNotificationService.kt
@@ -1,0 +1,61 @@
+package uk.gov.justice.digital.hmpps.hmppsinterventionsservice.service.notifications
+
+import mu.KLogging
+import org.springframework.beans.factory.annotation.Value
+import org.springframework.context.ApplicationListener
+import org.springframework.stereotype.Service
+import uk.gov.justice.digital.hmpps.hmppsinterventionsservice.component.EmailSender
+import uk.gov.justice.digital.hmpps.hmppsinterventionsservice.events.ReferralEvent
+import uk.gov.justice.digital.hmpps.hmppsinterventionsservice.events.ReferralEventType
+import uk.gov.justice.digital.hmpps.hmppsinterventionsservice.exception.AsyncEventExceptionHandling
+import uk.gov.justice.digital.hmpps.hmppsinterventionsservice.service.HMPPSAuthService
+import uk.gov.justice.digital.hmpps.hmppsinterventionsservice.service.NotifyService
+import javax.transaction.Transactional
+
+@Service
+@Transactional
+class ReferralNotificationService(
+  @Value("\${notify.templates.referral-sent}") private val referralSentTemplateID: String,
+  @Value("\${notify.templates.referral-assigned}") private val referralAssignedTemplateID: String,
+  @Value("\${interventions-ui.baseurl}") private val interventionsUIBaseURL: String,
+  @Value("\${interventions-ui.locations.service-provider.referral-details}") private val spReferralDetailsLocation: String,
+  private val emailSender: EmailSender,
+  private val hmppsAuthService: HMPPSAuthService,
+) : ApplicationListener<ReferralEvent>, NotifyService {
+
+  companion object : KLogging()
+
+  @AsyncEventExceptionHandling
+  override fun onApplicationEvent(event: ReferralEvent) {
+    when (event.type) {
+      ReferralEventType.SENT -> {
+        val location = generateResourceUrl(interventionsUIBaseURL, spReferralDetailsLocation, event.referral.id)
+        val intervention = event.referral.intervention
+        val serviceProvider = intervention.dynamicFrameworkContract.primeProvider
+        emailSender.sendEmail(
+          referralSentTemplateID,
+          intervention.incomingReferralDistributionEmail,
+          mapOf(
+            "organisationName" to serviceProvider.name,
+            "referenceNumber" to event.referral.referenceNumber!!,
+            "referralUrl" to location.toString(),
+          )
+        )
+      }
+
+      ReferralEventType.ASSIGNED -> {
+        val userDetails = hmppsAuthService.getUserDetail(event.referral.currentAssignee!!)
+        val location = generateResourceUrl(interventionsUIBaseURL, spReferralDetailsLocation, event.referral.id)
+        emailSender.sendEmail(
+          referralAssignedTemplateID,
+          userDetails.email,
+          mapOf(
+            "spFirstName" to userDetails.firstName,
+            "referenceNumber" to event.referral.referenceNumber!!,
+            "referralUrl" to location.toString(),
+          )
+        )
+      }
+    }
+  }
+}
